Move array length operation to top level class
